Introduced NLogBeginScopeParser that provides caching of property-reflection across all NLogLogger-objects (Reduce allocation and better cache reuse).
